Refactor perf observer
